added comments and changed samp to float
final...uhh yeah sure

+++ lab3/RTDSP/intio.c
@@ -146,11 +146,19 @@ void init_HWI(void)
 /******************** WRITE YOUR INTERRUPT SERVICE ROUTINE HERE***********************/  
 void ISR_AIC(void){
 
-	Int16 samp;
+	float samp;
 	//samp = mono_read_16Bit();				//read sample from codec. reads L & R sample from audio port and creates a mono average. returns 16bit integer
+	
+	/* sinegen outputs in range 0-1, from sine.c we have 
+	 * (!DSK6713_AIC23_write(H_Codec, ((Int32)(sample * L_Gain))))
+	 * where gain is Int32 L_Gain = 2100000000;
+	 * divide that by 2^32 -1, then multiply for a 16bit integer gives around 32,000 as a gain
+	 * to give a sensibly sized output
+	 * */
+	 
 	samp = sinegen()*32000;
 	samp = abs(samp);					//fullwave rectify function, take absolute value of the signal amplitude
-	mono_write_16Bit(samp);			//write out rectified value. nb samp < 16bits
+	mono_write_16Bit((Int16)samp);			//write out rectified value. nb samp < 16bits
